// src/payload-types.ts
export type OrderStatus = ('processing' | 'completed' | 'cancelled' | 'refunded') | null;
// ...
export interface User {
  id: number;
  name?: string | null;
  roles?: ('admin' | 'customer')[] | null;
  orders?: {
    docs?: (number | Order)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  cart?: {
    docs?: (number | Cart)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  addresses?: {
    docs?: (number | Address)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "orders".
 */
export interface Order {
  id: number;
  items?:
    | {
        product?: (number | null) | Product;
        variant?: (number | null) | Variant;
        quantity: number;
        id?: string | null;
      }[]
    | null;
  shippingAddress?: {
    title?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    company?: string | null;
    addressLine1?: string | null;
    addressLine2?: string | null;
    city?: string | null;
    state?: string | null;
    postalCode?: string | null;
    country?: string | null;
    phone?: string | null;
  };
  customer?: (number | null) | User;
  customerEmail?: string | null;
  transactions?: (number | Transaction)[] | null;
  status?: OrderStatus;
  amount?: number | null;
  currency?: 'MVR' | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "products".
 */
export interface Product {
  id: number;
  title: string;
  description?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  gallery?:
    | {
        image: number | Media;
        variantOption?: (number | null) | VariantOption;
        id?: string | null;
      }[]
    | null;
  layout?: (CallToActionBlock | ContentBlock | MediaBlock)[] | null;
  inventory?: number | null;
  enableVariants?: boolean | null;
  variantTypes?: (number | VariantType)[] | null;
  variants?: {
    docs?: (number | Variant)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  priceInMVREnabled?: boolean | null;
  priceInMVR?: number | null;
  relatedProducts?: (number | Product)[] | null;
  meta?: {
    title?: string | null;
    /**
     * Maximum upload file size: 12MB. Recommended file size for images is <500KB.
     */
    image?: (number | null) | Media;
    description?: string | null;
  };
  categories?: (number | Category)[] | null;
  /**
   * When enabled, the slug will auto-generate from the title field on save and autosave.
   */
  generateSlug?: boolean | null;
  slug: string;
  updatedAt: string;
  createdAt: string;
  deletedAt?: string | null;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  alt: string;
  caption?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "variantOptions".
 */
export interface VariantOption {
  id: number;
  _variantOptions_options_order?: string | null;
  variantType: number | VariantType;
  label: string;
  /**
   * should be defaulted or dynamic based on label
   */
  value: string;
  updatedAt: string;
  createdAt: string;
  deletedAt?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "variantTypes".
 */
export interface VariantType {
  id: number;
  label: string;
  name: string;
  options?: {
    docs?: (number | VariantOption)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  updatedAt: string;
  createdAt: string;
  deletedAt?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "CallToActionBlock".
 */
export interface CallToActionBlock {
  richText?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  links?:
    | {
        link: {
          type?: ('reference' | 'custom') | null;
          newTab?: boolean | null;
          reference?: {
            relationTo: 'pages';
            value: number | Page;
          } | null;
          url?: string | null;
          label: string;
          /**
           * Choose how the link should be rendered.
           */
          appearance?: ('default' | 'outline') | null;
        };
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'cta';
}
// ...
export interface ContentBlock {
  columns?:
    | {
        size?: ('oneThird' | 'half' | 'twoThirds' | 'full') | null;
        richText?: {
          root: {
            type: string;
            children: {
              type: any;
              version: number;
              [k: string]: unknown;
            }[];
            direction: ('ltr' | 'rtl') | null;
            format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
            indent: number;
            version: number;
          };
          [k: string]: unknown;
        } | null;
        enableLink?: boolean | null;
        link?: {
          type?: ('reference' | 'custom') | null;
          newTab?: boolean | null;
          reference?: {
            relationTo: 'pages';
            value: number | Page;
          } | null;
          url?: string | null;
          label: string;
          /**
           * Choose how the link should be rendered.
           */
          appearance?: ('default' | 'outline') | null;
        };
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'content';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "MediaBlock".
 */
export interface MediaBlock {
  media: number | Media;
  id?: string | null;
  blockName?: string | null;
  blockType: 'mediaBlock';
}
// ...
export interface Category {
  id: number;
  title: string;
  /**
   * When enabled, the slug will auto-generate from the title field on save and autosave.
   */
  generateSlug?: boolean | null;
  slug: string;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface Variant {
  id: number;
  /**
   * Used for administrative purposes, not shown to customers. This is populated by default.
   */
  title?: string | null;
  product: number | Product;
  options: (number | VariantOption)[];
  inventory?: number | null;
  priceInMVREnabled?: boolean | null;
  priceInMVR?: number | null;
  updatedAt: string;
  createdAt: string;
  deletedAt?: string | null;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "transactions".
 */
export interface Transaction {
  id: number;
  items?:
    | {
        product?: (number | null) | Product;
        variant?: (number | null) | Variant;
        quantity: number;
        id?: string | null;
      }[]
    | null;
  billingAddress?: {
    title?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    company?: string | null;
    addressLine1?: string | null;
    addressLine2?: string | null;
    city?: string | null;
    state?: string | null;
    postalCode?: string | null;
    country?: string | null;
    phone?: string | null;
  };
  status: 'pending' | 'succeeded' | 'failed' | 'cancelled' | 'expired' | 'refunded';
  customer?: (number | null) | User;
  customerEmail?: string | null;
  order?: (number | null) | Order;
  cart?: (number | null) | Cart;
  amount?: number | null;
  currency?: 'MVR' | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "carts".
 */
export interface Cart {
  id: number;
  items?:
    | {
        product?: (number | null) | Product;
        variant?: (number | null) | Variant;
        quantity: number;
        id?: string | null;
      }[]
    | null;
  customer?: (number | null) | User;
  purchasedAt?: string | null;
  status?: ('active' | 'purchased' | 'abandoned') | null;
  subtotal?: number | null;
  currency?: 'MVR' | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "addresses".
 */
export interface Address {
  id: number;
  customer?: (number | null) | User;
  title?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  company?: string | null;
  addressLine1?: string | null;
  addressLine2?: string | null;
  city?: string | null;
  state?: string | null;
  postalCode?: string | null;
  country:
    | 'US'
    | 'GB'
    | 'CA'
    | 'AU'
    | 'AT'
    | 'BE'
    | 'BR'
    | 'BG'
    | 'CY'
    | 'CZ'
    | 'DK'
    | 'EE'
    | 'FI'
    | 'FR'
    | 'DE'
    | 'GR'
    | 'HK'
    | 'HU'
    | 'IN'
    | 'IE'
    | 'IT'
    | 'JP'
    | 'LV'
    | 'LT'
    | 'LU'
    | 'MY'
    | 'MT'
    | 'MX'
    | 'NL'
    | 'NZ'
    | 'NO'
    | 'PL'
    | 'PT'
    | 'RO'
    | 'SG'
    | 'SK'
    | 'SI'
    | 'ES'
    | 'SE'
    | 'CH';
  phone?: string | null;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface Quotation {
  id: number;
  user: number | User;
  items?:
    | {
        product: number | Product;
        variant?: (number | null) | Variant;
        quantity: number;
        /**
         * Price of a single item at the time of quotation.
         */
        price?: number | null;
        id?: string | null;
      }[]
    | null;
  total: number;
  status?: ('draft' | 'sent' | 'expired') | null;
  updatedAt: string;
  createdAt: string;
}